Add authentication component for displaying user name when logged in
```
Added authentication component for displaying user name when logged in.

A new component, "Auth", has been added to the "src/components" directory. This component imports the "getSession" function from the "auth-astro/server" module and retrieves the session using "Astro.request". If a session exists, it displays a welcome message with the user's name, otherwise it shows a message indicating that the user is not logged in.
```